requirements.txt
opencv-python
face_recognition
dlib
Pillow
pyzbar
